requirements.txt
openai
langchain
python-dotenv
faiss-cpu
PyPDF2
requests